PDF creation does not need a baseURI
Resolving internal URIs is not needed within the PDF creation,
as rendered image can be included via absolute URI as well.

This makes configuration much easier, as no root URI as to
be provided which has been taken from the servlet context.
So, no BaseController has not to be ServletContextAware
anymore :)