Add beyond timespan values to dataset metadata
A very common use case for requesting dataset values for a
particular timespan is to render values in a chart. However,
rendering values stop at the bounds of the requested time
interval. A client has to know the next values beyond the
actual requested timespan to render the graph properly.

Within the expanded view of a data collection the `extras`
section contains now both the previous value beyond lower
and following value beyond upper interval bounds.

Closes:
  #440