Add next/previous value beyond requested interval
Clients may render charts according to the requested
time interval. However, to render until the actual
requested interval borders the next value beyond
have to be known.

Two additional select statements query the next value
beyond a timestamp. We do a sort (asc/desc) on the
timestamp column for each of the two requests. Also,
the output is limited to 1 so the that database can
do the optimization of the query.

As this feature is sharing TimeseriesDataMetadata with
reference values changed the way metadata is being set.