Move custom Geojson implementation to temp subfolder
The custom Geojson implementation was limited to Point geometries.
However, some use cases requires a Station geometry be more
flexible, i.e. to support other geometry types as well.

Custom Geojson implementation is too limited and it makes sense to
bet on existing and proven solutions. We are going to use Geojson
encoding as version 2.x introduced (and actually borrowed from
SOS project). Moving the existing code to subfolder is a prestep of
the following refactoring.