Allow to set a value formatter
Having output types holding generic values has the disadvantage that
values are not so easy to format when needed. This is not true for
quantity values in particular, but for those holding generic values
itself (container values, e.g. profile values or aggregates/records).

When creating a value in associated repository the formatter can be
created and set according to the value type. A custom ValueFormatter
serves as functional interface to be set directly on the value to be
formatted.