Always write multiple series as zip response
We want a CSV output to be consistently show meta information on top of
the CSV file, i.e. event before the actual column header starts. For
example:

    Phenomenon: Temperature
    Sensor: Thermometer
    ...
    col_1;col2;...

For multiple series entries where header entries differ, this have to
results in multiple files being written and zipped as single file for
download. For a single series entry, the CSV can directly be written
to the response stream.